Divided the start activity into two
