+++ app/src/main/AndroidManifest.xml
@@ -23,6 +23,15 @@
             </intent-filter>
         </activity>
         <activity
+            android:name=".activities.SpeedActivity"
+            android:screenOrientation="portrait">
+            <intent-filter>
+                <action android:name="android.intent.action.VIEW" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        <activity
             android:name=".activities.StartActivity"
             android:screenOrientation="portrait">
             <intent-filter>
